requirements.txt
aiogram==3.14.0
python-dotenv==1.0.1
opencv-python==4.9.0.80
pyzbar==0.1.9
Pillow==10.3.0
numpy==1.26.4
aiohttp==3.9.5
pytesseract==0.3.10